Replace old HttpSysConsoleClient
