+++ HttpSysConsoleClient/.vscode/tasks.json
@@ -7,7 +7,7 @@
             "command": "dotnet",
             "args": [
                 "build",
-                "${workspaceFolder}/HttpSysConsoleClient",
+                "${workspaceFolder}/ConsoleSystemBrowser",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
