--- HttpSysConsoleClient/ConsoleSystemBrowser/packages.config
@@ -1,46 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="IdentityModel" version="2.8.0" targetFramework="net452" />
-  <package id="IdentityModel.OidcClient" version="2.1.2" targetFramework="net46" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="1.1.0" targetFramework="net452" />
-  <package id="Microsoft.Extensions.Logging" version="1.1.1" targetFramework="net452" />
-  <package id="Microsoft.Extensions.Logging.Abstractions" version="1.1.1" targetFramework="net452" />
-  <package id="Microsoft.IdentityModel.Logging" version="1.1.3" targetFramework="net452" />
-  <package id="Microsoft.IdentityModel.Tokens" version="5.1.3" targetFramework="net452" />
-  <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="net452" />
-  <package id="NETStandard.Library" version="1.6.1" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net452" />
-  <package id="System.Collections" version="4.3.0" targetFramework="net452" />
-  <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net452" />
-  <package id="System.ComponentModel" version="4.3.0" targetFramework="net452" />
-  <package id="System.Diagnostics.Debug" version="4.3.0" targetFramework="net452" />
-  <package id="System.Diagnostics.Tools" version="4.3.0" targetFramework="net452" />
-  <package id="System.Diagnostics.Tracing" version="4.3.0" targetFramework="net452" />
-  <package id="System.Globalization" version="4.3.0" targetFramework="net452" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="5.1.3" targetFramework="net452" />
-  <package id="System.IO" version="4.3.0" targetFramework="net452" />
-  <package id="System.IO.Compression" version="4.3.0" targetFramework="net452" requireReinstallation="true" />
-  <package id="System.Linq" version="4.3.0" targetFramework="net452" />
-  <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net452" />
-  <package id="System.Net.Http" version="4.3.1" targetFramework="net452" requireReinstallation="true" />
-  <package id="System.Net.Primitives" version="4.3.0" targetFramework="net452" />
-  <package id="System.ObjectModel" version="4.3.0" targetFramework="net452" />
-  <package id="System.Reflection" version="4.3.0" targetFramework="net452" />
-  <package id="System.Reflection.Extensions" version="4.3.0" targetFramework="net452" />
-  <package id="System.Reflection.Primitives" version="4.3.0" targetFramework="net452" />
-  <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="net452" />
-  <package id="System.Runtime" version="4.3.0" targetFramework="net452" />
-  <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="net452" />
-  <package id="System.Runtime.InteropServices" version="4.3.0" targetFramework="net452" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net452" />
-  <package id="System.Runtime.Numerics" version="4.3.0" targetFramework="net452" />
-  <package id="System.Text.Encoding" version="4.3.0" targetFramework="net452" />
-  <package id="System.Text.Encoding.Extensions" version="4.3.0" targetFramework="net452" />
-  <package id="System.Text.RegularExpressions" version="4.3.0" targetFramework="net452" />
-  <package id="System.Threading" version="4.3.0" targetFramework="net452" />
-  <package id="System.Threading.Tasks" version="4.3.0" targetFramework="net452" />
-  <package id="System.Threading.Timer" version="4.3.0" targetFramework="net452" />
-  <package id="System.ValueTuple" version="4.3.0" targetFramework="net452" />
-  <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="net452" requireReinstallation="true" />
-  <package id="System.Xml.XDocument" version="4.3.0" targetFramework="net452" />
-</packages>
