--- HttpSysConsoleClient/HttpSysConsoleClient.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HttpSysConsoleClient", "HttpSysConsoleClient\HttpSysConsoleClient.csproj", "{FE9F6ECE-0250-4461-B77B-5038D525243A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FE9F6ECE-0250-4461-B77B-5038D525243A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FE9F6ECE-0250-4461-B77B-5038D525243A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FE9F6ECE-0250-4461-B77B-5038D525243A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FE9F6ECE-0250-4461-B77B-5038D525243A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
