--- HttpSysConsoleClient/HttpSysConsoleClient/HttpSysConsoleClient.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="IdentityModel.OidcClient" Version="6.0.0-preview.1" />
-  </ItemGroup>
-
-</Project>
